requirements.txt
python-telegram-bot==22.5
python-dotenv==1.0.0
aiohttp==3.9.1
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
gunicorn==21.2.0
supabase==2.24.0
websockets==15.0.1
